# Improve performance of quadtree point-to-polyline join 