@trxcllnt 2021-03-26
@thomcom yeah for sure. We might also want to wait on this PR because, while it's faster, it requires allocating much more temporary memory than the current implementation. I _think_ it's possible to apply some of the techniques in the hausdorff iterator to eliminate that extra memory, but I'd need to explore it in depth.