@trxcllnt 2021-04-07
@harrism I re-requested your review for the new _even better_ reimplementation of `quadtree_point_to_nearest_polyline.cu`