@harrism 2021-05-04
> @harrism I re-requested your review for the new version of `quadtree_point_to_nearest_polyline.cu`

I totally missed this. I'm sorry.