@cwharris 2021-05-20
I had questions regarding how the tests were updated due to the new sortedness of the outputs, and @trxcllnt answered those offline. lgtm.